refactor: merge DetailsComponent and OverviewComponent into ProjectComponent
