+++ modules/features/project/project/src/main/java/lu/aqu/projper/project/ProjectComponent.kt
@@ -1,17 +1,20 @@
-package lu.aqu.projper.project.overview.di
+package lu.aqu.projper.project
 
 import dagger.BindsInstance
 import dagger.Component
-import dagger.android.AndroidInjector
 import lu.aqu.core.di.FragmentScope
 import lu.aqu.core.di.Injector
-import lu.aqu.projper.project.ProjectModule
+import lu.aqu.projper.project.details.DetailsFragment
 import lu.aqu.projper.project.overview.OverviewFragment
 import retrofit2.Retrofit
 
 @Component(modules = [ProjectModule::class])
 @FragmentScope
-interface OverviewComponent : AndroidInjector<OverviewFragment>, Injector {
+interface ProjectComponent : Injector {
+
+    fun inject(overviewFragment: OverviewFragment)
+
+    fun inject(detailsFragment: DetailsFragment)
 
     @Component.Builder
     interface Builder {
@@ -19,6 +22,6 @@ interface OverviewComponent : AndroidInjector<OverviewFragment>, Injector {
         @BindsInstance
         fun retrofit(retrofit: Retrofit): Builder
 
-        fun build(): OverviewComponent
+        fun build(): ProjectComponent
     }
 }
