--- modules/features/project/project/src/main/java/lu/aqu/projper/project/details/di/DetailsComponent.kt
@@ -1,24 +1,0 @@
-package lu.aqu.projper.project.details.di
-
-import dagger.BindsInstance
-import dagger.Component
-import dagger.android.AndroidInjector
-import lu.aqu.core.di.FragmentScope
-import lu.aqu.core.di.Injector
-import lu.aqu.projper.project.ProjectModule
-import lu.aqu.projper.project.details.DetailsFragment
-import retrofit2.Retrofit
-
-@Component(modules = [ProjectModule::class])
-@FragmentScope
-interface DetailsComponent : AndroidInjector<DetailsFragment>, Injector {
-
-    @Component.Builder
-    interface Builder {
-
-        @BindsInstance
-        fun retrofit(retrofit: Retrofit): Builder
-
-        fun build(): DetailsComponent
-    }
-}
